fix(ci): make the test tree build on windows and linux, not just darwin
The 3-OS pre-tag matrix failed on Windows at `go vet ./...` with two
test-only build breaks, and fail-fast cancelled Linux before it could
report a third. All three are the same defect: a test reaches a
platform-specific API and is guarded (or not) at RUN time, when the
guard has to be at BUILD time.

internal/cli/watch_exit_6179_test.go
internal/cli/watch_exit_signal_6179_unix_test.go (new)
  TestRunWatch_SignalExitsSuccessfully used syscall.Kill, which does not
  exist on Windows. TAGGED (`//go:build !windows`), not made portable,
  and split into its own file so the other four tests in the original
  file keep running on all three platforms.

  The intent here really is unix signal semantics: the assertion is that
  SIGTERM is HANDLED and the handler returns normally (exit 0), because
  launchd/systemd treat death BY an unhandled signal as an unsuccessful
  exit and re-trigger KeepAlive/Restart — the reap-respawn oscillation
  #6179 fixed. Windows cannot express that. os.Process.Signal accepts
  only os.Kill there, which terminates outright and can never be observed
  by a handler, so the "portable" rewrite would assert the opposite of the
  contract. There is also nothing on Windows for it to protect: the
  respawn semantics belong to launchd and systemd, not schtasks. The
  reasoning is written into the new file's header.

internal/cli/watcher_fleet_6180_test.go
internal/cli/watcher_persist_6180_darwin_test.go (new)
  TestDarwinWatcherUnloadIsPersistent used the darwin-only seam
  watchers.SetLaunchctlRunnerForTest behind a `runtime.GOOS != "darwin"`
  t.Skip — which guarded the run but not the build, so the whole file
  failed to compile off macOS. This break was never reported by CI:
  vet stops at the first failure and had already died in internal/install.

  Moved to a `_darwin_test.go` file, matching this package's existing
  convention (watcher_ctl_detect_darwin_test.go). The now-redundant t.Skip
  is removed rather than left as dead code. The assertion is genuinely
  darwin-only — launchctl's disable/enable verbs and their ordering
  relative to bootstrap — with no systemd or schtasks equivalent, so
  nothing is lost off macOS. The rest of the file is untouched and stays
  cross-platform; the unused `runtime` import is dropped.

internal/install/watcher_write_nonfatal_r3_test.go
  MADE PORTABLE, not tagged. Nothing about
  TestApply_WatcherWriteFailureIsNonFatal is macOS-specific: the property
  is install.Apply's per-repo warn-and-continue, and watchers.Write
  rejects a control byte in Unit.Repo via validateUnitFields on every
  platform. It only failed to build because of two incidental spellings,
  and both portable replacements already exist in this package:

    - watchers.SetLaunchctlRunnerForTest -> watchers.StubServiceCallsForTest,
      the documented single cross-platform stub, which short-circuits
      launchctl, systemctl AND schtasks. launchctl_stub_darwin_test.go
      already delegates to it for exactly this reason.
    - the ad-hoc t.Setenv("HOME", ...) -> testsupport.IsolateHome, which
      also sets %USERPROFILE% (os.UserHomeDir reads that on Windows, and
      watchers.UnitDir goes through it) and asserts the redirect took
      effect. guidance_test.go documents the same trap.

  No assertion was weakened, skipped or deleted.

  This one was PRE-EXISTING AND ALREADY DOCUMENTED. #6197's PR body said
  verbatim that "internal/install's test binary does not compile off
  macOS — watcher_write_nonfatal_r3_test.go uses the darwin-only
  SetLaunchctlRunnerForTest with no build tag, so `GOOS=linux go vet
  ./internal/install/...` fails on origin/main too. No Linux contributor
  or CI runner can build that package's tests." It was written down as a
  known limit and left unfixed. It is now a release blocker.

Verification
  Cross-GOOS vet cannot be run in full on this machine: the tree-sitter
  bindings are cgo, so CGO_ENABLED=0 breaks them before type-checking
  anything downstream, and no cross C toolchain is installed. What was
  run:

    go vet ./...                                    darwin   -> 0
    go vet -tags perf ./...                         darwin   -> 0
    go build ./...                                  darwin   -> 0
    gofmt -l .                                      silent
    GOOS=windows CGO_ENABLED=0 go vet ./internal/install/...  -> 0
    GOOS=linux   CGO_ENABLED=0 go vet ./internal/install/...  -> 0
    GOOS=windows CGO_ENABLED=0 go vet ./...   -> only the cgo/tree-sitter
    GOOS=linux   CGO_ENABLED=0 go vet ./...      type errors remain; zero
                                                 findings outside them
    go test -count=1 ./internal/cli/... ./internal/install/...  -> 0

  internal/install is outside the cgo cone, so its cross-GOOS vet is a
  direct proof: it returns 0 on windows and linux where it failed on
  origin/main. internal/cli is inside the cone and could not be
  type-checked cross-GOOS; both files edited there are UNTAGGED, so
  darwin's vet type-checks the identical source the other platforms will
  see, and `go list` with GOOS set was used to prove the tagged files are
  actually excluded (windows and linux select neither the _unix_test.go
  nor the _darwin_test.go file; linux selects the former only; darwin
  selects both).

Sweep
  The whole tree was swept for the same class rather than only the named
  files: every file `go list` selects for GOOS=windows and GOOS=linux
  (with and without -tags perf, 4381 files) was grepped for unix-only and
  darwin-only symbols — syscall.Kill/Mkfifo/Umask/Getrlimit/Setpgid/
  Statfs/Wait4/Syscall/SYS_*/SIGUSR*/Credential, golang.org/x/sys/unix,
  SetLaunchctlRunnerForTest — and for windows-only API leaking into unix
  builds. After this change the only remaining hits are string literals
  (a regexp source in guard_6053_liveness_test.go, an error message in
  test_isolation_guard.go, a map key in extractors/golang/imports.go).
  Everything else was already correctly excluded by an existing
  `//go:build !windows` or `//go:build darwin`.
